+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # v4.5.1
+* Changed: Improve Accumulo iterator memory by not converting values to strings and then back again
 * Removed: Ability to run embedded Elasticsearch
 
 # v4.5.0